# SwaggerOffLine.yaml
openapi: 3.1.0
info:
  title: Auto Escola 3ESR
  description: "API REST para gerenciamento de instrutores, alunos, usuários e agendamento/cancelamento\
    \ de instruções da auto escola 3ESR."
  contact:
    name: Auto Escola 3ESR
    email: autoescola3esr@email.com.br
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: v1
servers:
- url: http://localhost:8081
  description: Generated server url
...
paths:
  /usuarios/{id}/perfil:
    put:
      tags:
      - Usuários
      summary: Atualizar perfil do usuário
      description: "Somente administradores. Perfis: ADMIN ou USER."
      operationId: atualizarPerfil
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoPerfil"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoUsuario"
      security:
      - bearer-key: []
  /usuarios/senha:
    put:
      tags:
      - Usuários
      summary: Alterar a própria senha
      description: Disponível para qualquer usuário autenticado.
      operationId: alterarPropriaSenha
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAlteracaoSenha"
        required: true
      responses:
        "204":
          description: Senha alterada
        "400":
          description: Senha atual incorreta ou nova senha inválida
      security:
      - bearer-key: []
  /instrutores:
    get:
      tags:
      - Instrutores
      summary: Listar instrutores ativos
      description: "Ordenado por nome (crescente), 10 registros por página."
      operationId: listarInstrutores
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "nome,ASC"
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDadosListagemInstrutor"
      security:
      - bearer-key: []
    put:
      tags:
      - Instrutores
      summary: Atualizar instrutor
      description: "Permite alterar apenas nome, telefone e endereço. E-mail, CNH\
        \ e especialidade não podem ser alterados."
      operationId: atualizarInstrutor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoInstrutor"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
      - bearer-key: []
    post:
      tags:
      - Instrutores
      summary: Cadastrar instrutor
      description: Disponível apenas para administradores.
      operationId: cadastrarInstrutor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroInstrutor"
        required: true
      responses:
        "201":
          description: Instrutor cadastrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
        "400":
          description: Dados inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
      - bearer-key: []
...
  /usuarios:
    get:
      tags:
      - Usuários
      summary: Listar usuários
      description: Somente administradores.
      operationId: listarUsuarios
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 10
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "login,ASC"
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageDadosDetalhamentoUsuario"
      security:
      - bearer-key: []
    post:
      tags:
      - Usuários
      summary: Cadastrar usuário
      description: Somente administradores. A senha é armazenada criptografada (BCrypt).
      operationId: cadastrarUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroUsuario"
        required: true
      responses:
        "201":
          description: Usuário cadastrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoUsuario"
        "403":
          description: Usuário sem perfil de administrador
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoUsuario"
      security:
      - bearer-key: []
  /login:
    post:
      tags:
      - Autenticação
      summary: Efetuar login
      description: "Retorna o token JWT a ser enviado no cabeçalho Authorization:\
        \ Bearer <token>."
      operationId: efetuarLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosLogin"
        required: true
      responses:
        "200":
          description: Login efetuado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosTokenJWT"
        "401":
          description: Login ou senha inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosTokenJWT"
...
  /usuarios/{id}:
    get:
      tags:
      - Usuários
      summary: Detalhar usuário
      description: Somente administradores.
      operationId: detalharUsuario
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoUsuario"
      security:
      - bearer-key: []
    delete:
      tags:
      - Usuários
      summary: Excluir usuário
      description: Somente administradores. Um administrador não pode excluir a si
        mesmo.
      operationId: excluirUsuario
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Usuário excluído
      security:
      - bearer-key: []
  /instrutores/{id}:
    get:
      tags:
      - Instrutores
      summary: Detalhar instrutor
      operationId: detalharInstrutor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Instrutor encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
        "404":
          description: Instrutor não encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
      - bearer-key: []
    delete:
      tags:
      - Instrutores
      summary: Excluir instrutor
      description: "Exclusão lógica: o instrutor é marcado como inativo."
      operationId: excluirInstrutor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Instrutor inativado
      security:
      - bearer-key: []
...
components:
  schemas:
    DadosAtualizacaoPerfil:
      type: object
      properties:
        perfil:
          type: string
          enum:
          - USER
          - ADMIN
      required:
      - perfil
    DadosDetalhamentoUsuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        perfil:
          type: string
          enum:
          - USER
          - ADMIN
    DadosAlteracaoSenha:
      type: object
      properties:
        senha_atual:
          type: string
          minLength: 1
        nova_senha:
          type: string
          maxLength: 2147483647
          minLength: 6
      required:
      - nova_senha
      - senha_atual
    DadosAtualizacaoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        telefone:
          type: string
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
        email:
          type: string
        cnh:
          type: string
        especialidade:
          type: string
          enum:
          - MOTOS
          - CARROS
          - VANS
          - CAMINHOES
      required:
      - id
    DadosEndereco:
      type: object
      properties:
        logradouro:
          type: string
          minLength: 1
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
          minLength: 1
        cidade:
          type: string
          minLength: 1
        uf:
          type: string
          minLength: 1
          pattern: "[A-Z]{2}"
        cep:
          type: string
          minLength: 1
          pattern: "[0-9]{5}-[0-9]{3}"
      required:
      - bairro
      - cep
      - cidade
      - logradouro
      - uf
    DadosDetalhamentoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        cnh:
          type: string
        especialidade:
          type: string
          enum:
          - MOTOS
          - CARROS
          - VANS
          - CAMINHOES
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
        ativo:
          type: boolean
...
    DadosCadastroUsuario:
      type: object
      properties:
        login:
          type: string
          maxLength: 100
          minLength: 0
        senha:
          type: string
          maxLength: 2147483647
          minLength: 6
        perfil:
          type: string
          enum:
          - USER
          - ADMIN
      required:
      - login
      - perfil
      - senha
    DadosLogin:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        senha:
          type: string
          minLength: 1
      required:
      - login
      - senha
    DadosTokenJWT:
      type: object
      properties:
        tokenJWT:
          type: string
    DadosCadastroInstrutor:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
        cnh:
          type: string
          minLength: 1
          pattern: "\\d{9,11}"
        especialidade:
          type: string
          enum:
          - MOTOS
          - CARROS
          - VANS
          - CAMINHOES
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
      required:
      - cnh
      - email
      - endereco
      - especialidade
      - nome
      - telefone
...
    PageDadosDetalhamentoUsuario:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosDetalhamentoUsuario"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        empty:
          type: boolean
        unsorted:
          type: boolean
    DadosListagemInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        cnh:
          type: string
        especialidade:
          type: string
          enum:
          - MOTOS
          - CARROS
          - VANS
          - CAMINHOES
    PageDadosListagemInstrutor:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosListagemInstrutor"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
  securitySchemes:
    bearer-key:
      type: http
      description: Informe o token JWT obtido em POST /login
      scheme: bearer
      bearerFormat: JWT