# src/main/java/br/com/fiap3esr/autoescola3esr/adapter/in/controller/response/usuario/DadosDetalhamentoUsuario.java
package br.com.fiap3esr.autoescola3esr.adapter.in.controller.response.usuario;

import br.com.fiap3esr.autoescola3esr.application.core.domain.Usuario;
import br.com.fiap3esr.autoescola3esr.shared.vo.enumeration.Role;

// A senha nunca é devolvida nas respostas da API
public record DadosDetalhamentoUsuario(
        Long id,
        String login,
        Role perfil) {
    public DadosDetalhamentoUsuario(Usuario usuario) {
        this(usuario.getId(), usuario.getLogin(), usuario.getPerfil());
    }
}